refractored: header broken into components
